requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.3.0=mkl
absl-py=0.11.0=pyhd3eb1b0_1
apscheduler=3.6.3=pypi_0
astor=0.8.1=py37_0
blas=1.0=mkl
c-ares=1.17.1=h27cfd23_0
ca-certificates=2021.1.19=h06a4308_0
certifi=2020.12.5=py37h06a4308_0
cffi=1.14.4=pypi_0
cryptography=3.3.1=pypi_0
cycler=0.10.0=pypi_0
decorator=4.4.2=pypi_0
gast=0.2.2=py37_0
google-pasta=0.2.0=py_0
grpcio=1.31.0=py37hf8bcb03_0
h5py=2.10.0=py37hd6299e0_1
hdf5=1.10.6=hb1b8bf9_0
importlib-metadata=2.0.0=py_1
intel-openmp=2020.2=254
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.0=py_1
kiwisolver=1.3.1=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libprotobuf=3.13.0.1=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
markdown=3.3.3=py37h06a4308_0
matplotlib=3.3.3=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.2.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
ncurses=6.2=he6710b0_1
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
opencv-python=4.5.1.48=pypi_0
openssl=1.1.1i=h27cfd23_0
opt_einsum=3.1.0=py_0
pandas=1.2.1=pypi_0
pillow=8.1.0=pypi_0
pip=20.3.3=py37h06a4308_0
protobuf=3.13.0.1=py37he6710b0_1
pycparser=2.20=pypi_0
pyparsing=2.4.7=pypi_0
python=3.7.9=h7579374_0
python-dateutil=2.8.1=pypi_0
python-telegram-bot=13.1=pypi_0
pytz=2020.5=pypi_0
readline=8.0=h7b6447c_0
scipy=1.5.2=py37h0b6359f_0
setuptools=51.3.3=py37h06a4308_4
sightseer=1.1.3=pypi_0
six=1.15.0=py37h06a4308_0
sqlite=3.33.0=h62c20be_0
tensorboard=1.15.0=pyhb230dea_0
tensorflow=1.15.0=mkl_py37h28c19af_0
tensorflow-base=1.15.0=mkl_py37he1670d9_0
tensorflow-estimator=1.15.1=pyh2649769_0
termcolor=1.1.0=py37_1
tk=8.6.10=hbc83047_0
tornado=6.1=pypi_0
tzlocal=2.1=pypi_0
webencodings=0.5.1=py37_1
werkzeug=0.16.1=py_0
wget=3.2=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
wrapt=1.12.1=py37h7b6447c_1
xz=5.2.5=h7b6447c_0
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
